calculation Function

+++ Hotel-backend/Common/Dto/RoomAllocationDto.cs
@@ -18,7 +18,7 @@ public class RoomAllocationDto
     public int ActualDemand { get; set; }
     public int RoomsSold { get; set; }
     public bool Confirmed { get; set; }
-    public int Revenue { get; set; }
+    public decimal Revenue { get; set; }
     public int QuarterForecast { get; set; }
 }
